Make parse_system_time_from_str return an Option
In the future we not only want to expose functions for deserializing
data using a serde Deserializer, but also expose the underlying string
parsing functionality.
This change marks a first step moving us towards that goal, in that we
change the return value of the parse_system_time_from_str function from
a deserializer specific error to a more generically applicable Option.