Introduce public parse_system_time_from_str function
Clients of this crate may not only want to deserialize an object
containing a SystemTime using serde, but may also merely want to parse a
string into a SystemTime object.
So far we have only exposed the serde compatible functions, but with
this change we expose the parse_system_time_from_str function which
works on a bare-bones string object.