Introduce TimeZone trait
Unfortunately, it turns out that we cannot completely ignore time zones,
because we have use cases where the data we receive is provided in the
form of seconds since 1970-01-01 in Eastern Standard Time (EST). Whoever
thought such an encoding is sane deserves to be beheaded, but we
ultimately are forced to be able to deal with this format.
As a first step towards doing so, this change introduces a very
primitive time zone mechanism: a trait that contains a constant that
acts as the offset from UTC.