Remove Result return from test functions
So far we have let test functions return a Result and let Rust's unit
testing framework report the result. This approach has the downside of
potentially obfuscating the actual source of the error, because without
additional context it may be hard to trace back which of the paths an
error was reported on.
To fix this problem, this change removes the Result return type from all
those functions, in favor of just unwrapping directly. That not only
fixes the aforementioned issue, but also is more concise.