Feature gate all dependencies
We are in a bit of a predicament as to how best to organize all the
timing related functionality we want to reuse. With the current approach
we basically put everything into this very crate (in contrast to having
multiple smaller crates).
In order to still have good control over dependencies and only use what
is truly required, this change introduces feature switches that guard
individual blocks of functionality.