Use chrono-tz for time-zone aware conversions
So far we were under the assumption that we would only need to
deserialize data that has a fixed time zone offset. That is, for US
Eastern time we were working with a fixed 5h offset. As it turned out,
the data source we care about actually reports data with daylight
savings time adjustments in effect. As such, the offset is no longer
fixed.
To mitigate this problem, this change pulls in chrono-tz, which properly
supports time-zones including daylight savings time changes. Right now
we do not care about anything but US Eastern time and we do not want to
expose chrono-tz internals to clients. Thus, we rename the time-zone
aware conversion functions to directly reflect the fact that they
assume New York based time stamps.