Support serializing a time as RFC 3339 time stamp with nanosecond resolution
By default we serialize (and print) times in RFC 3339 format with
millisecond resolution, as that is a format that is friendly for the eye
while also having "reasonable" resolution. There are cases, however,
where a user may want the full nanosecond resolution that Rust uses
internally.
To enable such uses, this change adds variant of the print &
serialization functions that emit the time stamp string with nanosecond
resolution.