Fix documentation generation on docs.rs
The generation of the documentation for docs.rs is failing because a
build is attempted without any features enabled -- something that we do
not allow.
With this change we explicitly list all the crate's features in the
corresponding Cargo.toml meta data to cause the docs.rs build to happen
with all of them enabled.